requirements.txt
streamlit==1.28.0
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.1
PyJWT==2.8.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
faker==20.1.0

